+++ worker/src/RTC/Router.cpp
@@ -668,7 +668,7 @@ namespace RTC
 		for (auto* consumer : consumers)
 		{
 			// Update MID RTP extension value.
-			auto& mid = consumer->GetRtpParameters().mid;
+			const auto& mid = consumer->GetRtpParameters().mid;
 
 			if (!mid.empty())
 				packet->UpdateMid(mid);
@@ -736,7 +736,7 @@ namespace RTC
 		this->mapConsumerProducer[consumer] = producer;
 
 		// Get all streams in the Producer and provide the Consumer with them.
-		for (auto& kv : producer->GetRtpStreams())
+		for (const auto& kv : producer->GetRtpStreams())
 		{
 			auto* rtpStream     = kv.first;
 			uint32_t mappedSsrc = kv.second;