+++ worker/src/RTC/SctpListener.cpp
@@ -19,7 +19,7 @@ namespace RTC
 		auto jsonStreamIdTableIt = jsonObject.find("streamIdTable");
 
 		// Add streamIdTable.
-		for (auto& kv : this->streamIdTable)
+		for (const auto& kv : this->streamIdTable)
 		{
 			auto streamId      = kv.first;
 			auto* dataProducer = kv.second;